requirements.txt
pandas
yfinance
sqlalchemy
python-dotenv
